Restructured a little.
