+++ cw03/zad2/include/util.h
